requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
gradio>=5.0.0
httpx>=0.27.0
numpy>=1.26.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
diskcache>=5.6.0
